Use same Various.cpp for win32 and netcore targets.
