+++ targets/win32/nanoCLR/PAL/blockstorageList_stubs.cpp
@@ -16,6 +16,11 @@ static BlockStorageDevice *s_primaryDevice = NULL;
 // void BlockStorageList::Initialize()
 //{
 //}
+void BlockStorageList_Initialize()
+{
+
+}
+
 //
 // bool BlockStorageList::InitializeDevices()
 //{
